refactor: 가독성과 구조단순화를 위한 파일명수정
